Implement Location and Role management APIs

+++ EMS-PRO-Backend/src/main/java/com/example/Ems_Pro/Repository/LocationRepositry.java
@@ -3,6 +3,6 @@ package com.example.Ems_Pro.Repository;
 import com.example.Ems_Pro.Entity.LocationEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationRepositry extends JpaRepository<LocationEntity,Integer> {
+public interface LocationRepositry extends JpaRepository<LocationEntity,Long> {
 
 }
